refactor module category
- move dtos to service directory
- update returned type is a pointer
