--- modules/category/infras/controller/http-gin/bulk_create_category_api.go
@@ -1,25 +1,0 @@
-package httpgin
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-	categorymodel "github.com/ntttrang/go-food-delivery-backend-service/modules/category/model"
-	"github.com/ntttrang/go-food-delivery-backend-service/shared/datatype"
-)
-
-func (ctrl *CategoryHttpController) CreateBulkCategoryAPI(c *gin.Context) {
-	var requestBodyData []categorymodel.CategoryInsertDto
-
-	if err := c.ShouldBindJSON(&requestBodyData); err != nil {
-		panic(datatype.ErrBadRequest.WithError(err.Error()))
-	}
-
-	// call business logic in service
-	ids, err := ctrl.bulkCreateCmdHdl.Execute(c.Request.Context(), requestBodyData)
-	if err != nil {
-		panic(err)
-	}
-
-	c.JSON(http.StatusCreated, gin.H{"data": gin.H{"ids": ids}})
-}
