+++ modules/category/infras/controller/http-gin/create_category_api.go
@@ -4,12 +4,12 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	categorymodel "github.com/ntttrang/go-food-delivery-backend-service/modules/category/model"
+	"github.com/ntttrang/go-food-delivery-backend-service/modules/category/service"
 	"github.com/ntttrang/go-food-delivery-backend-service/shared/datatype"
 )
 
 func (ctrl *CategoryHttpController) CreateCategoryAPI(c *gin.Context) {
-	var requestBodyData categorymodel.CategoryInsertDto
+	var requestBodyData service.CategoryInsertDto
 
 	if err := c.ShouldBindJSON(&requestBodyData); err != nil {
 		panic(datatype.ErrBadRequest.WithError(err.Error()))
