+++ modules/category/infras/repository/gorm-mysql/delete_category.go
@@ -1,4 +1,4 @@
-package repository
+package categorygormmysql
 
 import (
 	"context"
