+++ modules/category/infras/repository/gorm-mysql/find_by_ids.go
@@ -1,4 +1,4 @@
-package repository
+package categorygormmysql
 
 import (
 	"context"
