+++ modules/category/infras/repository/gorm-mysql/update_category_by_id.go
@@ -1,14 +1,14 @@
-package repository
+package categorygormmysql
 
 import (
 	"context"
 
 	"github.com/google/uuid"
-	categorymodel "github.com/ntttrang/go-food-delivery-backend-service/modules/category/model"
+	"github.com/ntttrang/go-food-delivery-backend-service/modules/category/service"
 	"github.com/pkg/errors"
 )
 
-func (r *CategoryRepo) Update(ctx context.Context, id uuid.UUID, dto categorymodel.CategoryUpdateReq) error {
+func (r *CategoryRepo) Update(ctx context.Context, id uuid.UUID, dto service.CategoryUpdateReq) error {
 	db := r.dbCtx.GetMainConnection().Begin()
 
 	if err := db.Table(dto.TableName()).Where("id = ?", id).Updates(dto).Error; err != nil {
