--- modules/category/model/category_delete_req.go
@@ -1,7 +1,0 @@
-package categorymodel
-
-import "github.com/google/uuid"
-
-type CategoryDeleteReq struct {
-	Id uuid.UUID
-}
