--- modules/category/model/category_detail_req.go
@@ -1,7 +1,0 @@
-package categorymodel
-
-import "github.com/google/uuid"
-
-type CategoryDetailReq struct {
-	Id uuid.UUID
-}
