--- modules/category/model/category_detail_res.go
@@ -1,5 +1,0 @@
-package categorymodel
-
-type CategoryDetailRes struct {
-	Category
-}
