--- modules/category/model/list_category_req.go
@@ -1,11 +1,0 @@
-package categorymodel
-
-import (
-	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
-)
-
-type ListCategoryReq struct {
-	SearchCategoryDto
-	sharedModel.PagingDto
-	sharedModel.SortingDto
-}
