--- modules/category/model/list_category_res.go
@@ -1,14 +1,0 @@
-package categorymodel
-
-import (
-	"github.com/google/uuid"
-	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
-)
-
-type ListCategoryRes struct {
-	Id          uuid.UUID `json:"id"`
-	Name        string    `json:"name"`
-	Description string    `json:"description"`
-	Status      string    `json:"status"`
-	sharedModel.DateDto
-}
