--- modules/category/model/search_category_dto.go
@@ -1,6 +1,0 @@
-package categorymodel
-
-type SearchCategoryDto struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
