+++ modules/category/service/create_category.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"strings"
 
 	"github.com/google/uuid"
 	categorymodel "github.com/ntttrang/go-food-delivery-backend-service/modules/category/model"
@@ -9,6 +10,32 @@ import (
 	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
 )
 
+// Define DTOs & validate
+type CategoryInsertDto struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+
+	Id uuid.UUID `json:"-"`
+}
+
+func (c *CategoryInsertDto) Validate() error {
+	c.Name = strings.TrimSpace(c.Name)
+
+	if c.Name == "" {
+		return categorymodel.ErrNameRequired
+	}
+
+	return nil
+}
+
+func (c CategoryInsertDto) ConvertToCategory() *categorymodel.Category {
+	return &categorymodel.Category{
+		Name:        c.Name,
+		Description: c.Description,
+	}
+}
+
+// Initilize service
 type ICreateRepo interface {
 	Insert(ctx context.Context, data *categorymodel.Category) error
 }
@@ -21,7 +48,8 @@ func NewCreateCommandHandler(catRepo ICreateRepo) *CreateCommandHandler {
 	return &CreateCommandHandler{catRepo: catRepo}
 }
 
-func (s *CreateCommandHandler) Execute(ctx context.Context, data *categorymodel.CategoryInsertDto) error {
+// Implement
+func (s *CreateCommandHandler) Execute(ctx context.Context, data *CategoryInsertDto) error {
 	if err := data.Validate(); err != nil {
 		return datatype.ErrBadRequest.WithWrap(err).WithDebug(err.Error())
 	}
