+++ modules/category/service/get_category_detail.go
@@ -10,6 +10,16 @@ import (
 	sharedModel "github.com/ntttrang/go-food-delivery-backend-service/shared/model"
 )
 
+// Define DTOs & validate
+type CategoryDetailReq struct {
+	Id uuid.UUID
+}
+
+type CategoryDetailRes struct {
+	categorymodel.Category
+}
+
+// Initilize service
 type IGetDetailRepo interface {
 	FindById(ctx context.Context, id uuid.UUID) (categorymodel.Category, error)
 }
@@ -22,19 +32,20 @@ func NewGetDetailQueryHandler(repo IGetDetailRepo) *GetDetailQueryHandler {
 	return &GetDetailQueryHandler{repo: repo}
 }
 
-func (hdl *GetDetailQueryHandler) Execute(ctx context.Context, req categorymodel.CategoryDetailReq) (categorymodel.CategoryDetailRes, error) {
+// Implement
+func (hdl *GetDetailQueryHandler) Execute(ctx context.Context, req CategoryDetailReq) (*CategoryDetailRes, error) {
 	category, err := hdl.repo.FindById(ctx, req.Id)
 
 	if err != nil {
 		if errors.Is(err, categorymodel.ErrCategoryNotFound) {
-			return categorymodel.CategoryDetailRes{}, datatype.ErrNotFound.WithDebug(categorymodel.ErrCategoryNotFound.Error())
+			return nil, datatype.ErrNotFound.WithDebug(categorymodel.ErrCategoryNotFound.Error())
 		}
-		return categorymodel.CategoryDetailRes{}, datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
+		return nil, datatype.ErrInternalServerError.WithWrap(err).WithDebug(err.Error())
 	}
 
 	if category.Status == sharedModel.StatusDelete {
-		return categorymodel.CategoryDetailRes{}, datatype.ErrDeleted.WithError(categorymodel.ErrCategoryIsDeleted.Error())
+		return nil, datatype.ErrDeleted.WithError(categorymodel.ErrCategoryIsDeleted.Error())
 	}
 
-	return categorymodel.CategoryDetailRes{Category: category}, nil
+	return &CategoryDetailRes{Category: category}, nil
 }
